PR-15 Create AppModalBottomSheet composable and use it
